+++ rollup.config.js
@@ -1,13 +1,11 @@
 'use strict'
 
-const flatten = require('flatten')
-const glob = require('glob')
 const typescript = require('@rollup/plugin-typescript')
 const tsConfig = require('./tsconfig.json')
 
 module.exports = {
-    input: flatten(tsConfig.include.map(pattern => glob.sync(pattern)))
-        .filter(filename => !filename.endsWith('.d.ts')),
+    // We're running rollup via gulp, so globs are OK.
+    input: ['./assets/scripts/*.ts'],
     output: {
         dir: tsConfig.compilerOptions.outDir,
         format: 'iife',