+++ assets/js/core.ts
@@ -1,0 +1,8 @@
+/// <reference path='./core/module.d.ts' />
+
+import { Base } from './core/base'
+import { BodyClass } from './core/body-class'
+
+const cores = [new BodyClass(jQuery)]
+
+jQuery.each(cores, (_, core: Base): void => core.run())