+++ assets/js/menu/module.d.ts
@@ -1,0 +1,5 @@
+declare const jentilMenuL10n: JentilMenuL10n
+
+interface JentilMenuL10n {
+    submenu: string
+}