+++ assets/scripts/menu/module.d.ts
@@ -1,6 +1,0 @@
-declare const jentilMenuL10n: JentilMenuL10n
-
-interface JentilMenuL10n
-{
-    submenu: string
-}