+++ bs-config-sample.js
@@ -1,8 +1,8 @@
 module.exports = {
     // browser: ['google chrome', 'firefox'],
     files: [
-        './dist/styles/**/*',
-        './dist/scripts/**/*',
+        './dist/css/**/*',
+        './dist/js/**/*',
         './app/partials/**/*',
         './app/templates/**/*',
         './app/libraries/Jentil/Setups/Views/**/*'
@@ -11,7 +11,7 @@ module.exports = {
     // notify: true,
     // open: false,
     // port: 3000,
-    proxy: 'localhost',
+    proxy: 'http://wordpress.localhost:8080',
     reloadDebounce: 3000
     // reloadDelay: 0,
 }