+++ assets/js/menu/menu-button.ts
@@ -13,12 +13,23 @@ export class MenuButton extends Base {
             (event: JQuery.ClickEvent): void => {
                 const jmenu = this._j('.js-main-menu')
 
-                jmenu.slideToggle(this._fx_duration, (): void => {
-                    jmenu.toggleClass('show hide').css({display: ''})
-                })
-
+                this.toggleIcon(jmenu)
+                this.toggleMenu(jmenu)
                 event.preventDefault()
             }
         )
+    }
+
+    private toggleIcon(menu: JQuery<EventTarget>): void {
+        const icon = menu.hasClass('hide') ? 'fas fa-times' : 'fas fa-bars'
+
+        this._j('.js-menu-button-icon')
+            .html(`<span class="${icon}" aria-hidden="true"></span>`)
+    }
+
+    private toggleMenu(menu: JQuery<EventTarget>) {
+        menu.slideToggle(this._fx_duration, (): void => {
+            menu.toggleClass('show hide').css({display: ''})
+        })
     }
 }