+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
-and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
+
+## [Unreleased] - 
+
+### Changed
+- Replace SCSS with [Tailwind CSS](https://tailwindcss.com)
+- Replace gulp with [Laravel Mix](https://laravel-mix.com)
 
 ## [0.11.3] - 2023-04-19