+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Load `privacypolicy` templates from `app/templates` directory
 
+### Removed
+- Remove `@grottopress/scss` NPM package
+
 ## [0.12.1] - 2023-04-21
 
 ### Changed