+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Remove `fa-sm` sizing from Fontawesome icons
 - Rewrite JS for menus
+- Limit menu JS enhancement to primary menu
 
 ### Removed
 - Remove `@grottopress/scss` NPM package